Finished all end-points

+++ FitnessBeheerDomain/Model/RunningSession.cs
@@ -3,8 +3,8 @@ public class RunningSession
 {
     public int Id { get; set; }
     public DateTime Date { get; set; }
-    public int Duration { get; set; }
-    public int AvgSpeed { get; set; }
+    public double Duration { get; set; }
+    public double AvgSpeed { get; set; }
     public List<RunningSessionDetail> Details { get; set; } = new();
 }
