+++ FitnessDL/Model/CyclingSessionEF.cs
@@ -12,11 +12,11 @@ public class CyclingSessionEF
     [Key]
     public int Id { get; set; }
     public DateTime Date { get; set; }
-    public int Duration { get; set; }
-    public int Avg_Watt { get; set; }
-    public int Max_Watt { get; set; }
-    public int Avg_Cadence { get; set; }
-    public int Max_Cadence { get; set; }
+    public double Duration { get; set; }
+    public double Avg_Watt { get; set; }
+    public double Max_Watt { get; set; }
+    public double Avg_Cadence { get; set; }
+    public double Max_Cadence { get; set; }
     public string TrainingType { get; set; }
     public string? Comment { get; set; }
     public MemberEF Member { get; set; }
